Extend the ppg sizes.

+++ lib/Collector/Collector.h
@@ -17,8 +17,8 @@ public:
 
     int getLastData(Batch& batch);
 private:
-    uint16_t redBuffer[BUFSIZE];
-    uint16_t irBuffer[BUFSIZE];
+    uint32_t redBuffer[BUFSIZE];
+    uint32_t irBuffer[BUFSIZE];
     int idx;
     MAX30105 pulseoximiter;
     // other classes
